gitignore file extended by .pyc files; card class descdriptions moved to a separate supporting module
